Godrej Properties Worli Mumbai
Godrej Project in Worli, Mumbai is spread across 5 acres of the land region. It is encircled by gorgeous green spaces everywhere. The homes have been created for incredible style and solace. The housing setups come in the contributions of 2, 3, and 4 BHK. The housing designs are made well and there is a ton of greenery around. There is a lot of natural air and adequate daylight coursing through. The balconies are wide with enormous windows to permit a lot of light and air to go into the houses.

For More Info:
Visit: godrej.builderproject.in